Remove the template parameter
